Make the apply_extra extra any Debian packages
There is no real advantage in duplicating the names of the Debian
packages that are fetched as extra data in the apply_extra script.
This will prevent issues like updating/adding a Debian package as
extra data to the manifest and forgetting to add it to the
apply_extra script.

https://phabricator.endlessm.com/T18501