Fix opening for non-local files
The field code '%f' we passed to the exec key in the desktop file should only be used
for a single file name and non-local files will be copied to a tmp place first. Change
this to '%U' instead, to allow for non-local files.

Part of https://phabricator.endlessm.com/T18843